+++ example/ios/Runner.swift
@@ -1,0 +1,11 @@
+//
+//  Runner.swift
+//  Runner
+//
+//  Created by Andrés Boedo on 9/18/20.
+//  Copyright © 2020 The Chromium Authors. All rights reserved.
+//
+
+import Foundation
+
+// this implementation is empty, but the file is needed for Swift to be imported.