+++ lib/errors.dart
@@ -27,7 +27,6 @@ enum PurchasesErrorCode {
 
 /// Helper to convert from PlatformExceptions to PurchasesErrorCodes
 class PurchasesErrorHelper {
-
   /// Use this to convert a PlatformException to a PurchasesErrorCode.
   /// It will return `PurchasesErrorCode.unknownError` if the error code is not
   /// in the range of PurchasesErrorCodes.
@@ -57,5 +56,4 @@ class PurchasesErrorHelper {
     }
     return PurchasesErrorCode.values[errorCode];
   }
-
 }